Prevent negative human score from turning positive post infection
The frag counter, usually on the bottom-left of the screen, is used
in Mush Match to count "contribution score". Certain actions will
affect a player's score depending on whether they benefit or hurt
that player's team.

When a player starts as a human, all contribution to their team is,
conversely, also a detriment to the mush team. Therefore, when they
are infected by another mush, default behaviour has this score become
negative to represent this detriment, which the player can make for
in order to have the contribution score equal to a newly spawned
mush player.

(Note: the only real purpose of the contribution score is for a sort of
measure of performance at Mush Match between players, as those who are
more beneficial to the team they either won or died as, will rank
higher in the match end scoreboard.)

This small change makes it so that a human with negative score, which
can be achieved through means such as teamkilling (which is a very bad
thing to do in Mush Match), will not have this score "re-negated" into
a positive score when they are infected.

What this means is that humans that kill other humans, either through
sabotage or (more commonly) from misjudging a fellow human as a mush,
upon becoming a mush player won't be seen as players with a higher
score; that is, this scenario is now, by default, an exception to the
previously described mechanism.

Signed-off-by: Gustavo Ramos Rehermann <rehermann6046@gmail.com>